// R3 APP/types.ts
export interface AuxiliarData {
    id: string;
    Cuenta: string;
    Tercero: string;
    NIT: string;
    Fecha: string; // DD/MM/YYYY
    Nota: string;
    DocNum: string;
    Debitos: number;
    Creditos: number;
    isModified?: boolean;
    comentario?: string;
}

export interface RetencionBaseData {
    id: string;
    Codigo: string;
    CuentaDocumento: string;
    Fecha: string;
    Tercero: string;
    Valor: number;
    Porcentaje: number;
    Base: number;
    comentario?: string;
    isModified?: boolean;
}

export interface DianData {
    id: string;
    TipoDeDocumento: string;
    Folio: string;
    Prefijo: string;
    DocumentoDIAN: string; // Prefijo + Folio
    Fecha: string; // DD/MM/YYYY
    FechaObj: Date | null;
    NITEMISOR: string;
    NombreEmisor: string;
    NITReceptor: string;
    NombreReceptor: string;
    IVA: number;
    Base: number;
    Total: number;
    Grupo: string;
    
    // Additional tax fields
    ICA?: number;
    IC?: number;
    INC?: number;
    Timbre?: number;
    INCBolsas?: number;
    INCarbono?: number;
    INCombustibles?: number;
    ICDatos?: number;
    ICL?: number;
    INPP?: number;
    IBUA?: number;
    ICUI?: number;
    ReteIVA?: number;
    ReteRenta?: number;
    ReteICA?: number;
}

export interface VentasComprasData {
    id: string;
    Documento: string;
    Fecha: string;
    FechaObj: Date | null;
    Cliente: string;
    NIT: string;
    VentaNeta: number;
    IVA: number;
    Dias: string;
    Descuento: number;
    RetIVA: number;
    RetICA: number;
    RetFuente: number;
    ValorTotal: number;
}
// ...
export interface ValidationData {
    nombreValidado: string;
    estado: string;
    dv: string;
}